requirements.txt
aiofiles==24.1.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.11.0
arcosparse==0.4.2
asgiref==3.10.0
attrs==25.3.0
basemap @ file:///D:/Code%20et%20D%C3%A9veloppement/API%20App%20OceanNaviSense-AI/copernicus_wmts_viewer/basemap-2.0.0-cp312-cp312-win_amd64.whl#sha256=3d26528568e86b46bae3ab04435ddf0a88753b7fe0040a14057b41d8a023ae4f
basemap_data==2.0.0
bleach==6.2.0
blinker==1.9.0
bokeh==3.8.0
boto3==1.40.45
botocore==1.40.45
branca==0.8.1
Brotli==1.1.0
cachetools==6.2.0
Cartopy==0.25.0
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
cloudpickle==3.1.1
colorama==0.4.6
colorcet==3.1.0
coloredlogs==15.0.1
contourpy==1.3.3
copernicusmarine==2.2.2
crc32c==2.7.1
cycler==0.12.1
dask==2025.9.1
Django==5.2.7
docutils==0.21.2
donfig==0.8.1.post1
fastapi==0.119.0
ffmpy==0.6.3
filelock==3.20.0
Flask==3.1.2
flatbuffers==25.9.23
folium==0.20.0
fonttools==4.60.1
fsspec==2025.9.0
GDAL @ file:///D:/Code%20et%20Developpement/API%20App%20OceanNaviSense-AI/copernicus_wmts_viewer/gdal-3.11.1-cp312-cp312-win_amd64.whl#sha256=e328c144311ae5dda09805a9fd77517ed5568177a993e23aa84af05b37e140bc
geographiclib==2.1
geopandas==1.1.1
geopy==2.4.1
geoviews==1.14.1
gitdb==4.0.12
GitPython==3.1.45
groovy==0.1.2
h11==0.16.0
h5netcdf==1.6.4
h5py==3.14.0
holoviews==1.21.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.3
humanfriendly==10.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
linkify-it-py==2.0.3
locket==1.0.0
lxml==6.0.2
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.6
mdit-py-plugins==0.5.0
mdurl==0.1.2
ml_dtypes==0.5.3
mpmath==1.3.0
narwhals==2.6.0
numcodecs==0.16.3
numpy==2.3.3
orjson==3.11.3
OWSLib==0.34.1
packaging==25.0
pandas==2.3.3
panel==1.8.1
param==2.2.1
partd==1.4.2
pillow==11.3.0
plotly==6.3.1
protobuf==6.32.1
pyarrow==21.0.0
pycountry==24.6.1
pydantic==2.11.10
pydantic_core==2.33.2
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.2
pyogrio==0.11.1
pyparsing==3.2.5
pyproj==3.7.2
pyreadline3==3.5.4
pyshp==2.3.1
pystac==1.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
pyviz_comms==3.0.6
PyYAML==6.0.3
referencing==0.36.2
requests==2.32.5
rich==14.2.0
rpds-py==0.27.1
ruff==0.14.0
s3transfer==0.14.0
safehttpx==0.1.6
semantic-version==2.10.0
semver==3.0.4
setuptools==80.9.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.48.0
streamlit==1.50.0
streamlit-folium==0.25.3
sympy==1.14.0
tenacity==9.1.2
toml==0.10.2
tomlkit==0.13.3
toolz==1.0.0
tornado==6.5.2
tqdm==4.67.1
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uc-micro-py==1.0.3
urllib3==2.5.0
uvicorn==0.37.0
watchdog==6.0.0
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.1.3
xarray==2025.9.1
xyzservices==2025.4.0
zarr==3.1.3
